ITMO_team/requirements.txt
asgiref==3.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.10
django-cleanup==9.0.0
EditorConfig==0.17.1
idna==3.10
jsbeautifier==1.15.4
lxml==6.0.2
peepdf-3==5.1.1
pillow==11.3.0
prettytable==3.16.0
ProgramExecutableAnalyzer==1.1.2
purple==1.0.0
pypdf==6.1.0
python-dotenv==1.2.1
pythonaes==1.0
requests==2.32.5
six==1.17.0
sorl-thumbnail==13.0.0
sqlparse==0.5.5
stpyv8==13.1.201.22
tzdata==2025.3
urllib3==2.5.0
wcwidth==0.2.14
